+++ misk-core/build.gradle
@@ -6,6 +6,7 @@ buildscript {
 
 dependencies {
   api project(':wisp-resource-loader')
+  api project(':wisp-ssl')
   implementation dep.bouncycastle
   implementation dep.guice
   implementation dep.guiceMultibindings
