+++ misk-core/src/main/kotlin/misk/security/ssl/TrustStore.kt
@@ -1,6 +1,4 @@
 package misk.security.ssl
 
-import java.security.KeyStore
-
 /** A set of trusted root certificates. */
-data class TrustStore(val keyStore: KeyStore)
+typealias TrustStore = wisp.security.ssl.TrustStore
