+++ misk-core/src/main/kotlin/misk/security/ssl/TrustStoreConfig.kt
@@ -2,9 +2,12 @@ package misk.security.ssl
 
 import misk.security.ssl.SslLoader.Companion.FORMAT_JCEKS
 import javax.inject.Inject
+import wisp.security.ssl.TrustStoreConfig as WispTrustStoreConfig
 
 data class TrustStoreConfig @Inject constructor(
   val resource: String,
   val passphrase: String? = null,
   val format: String = FORMAT_JCEKS
-)
+) {
+  internal fun toWispConfig() = WispTrustStoreConfig(resource, passphrase, format)
+}
