--- misk-core/src/test/resources/ssl/client_cert.pem
@@ -1,21 +1,0 @@
------BEGIN CERTIFICATE-----
-MIIDgDCCAmigAwIBAgIEARuoZDANBgkqhkiG9w0BAQsFADBoMQswCQYDVQQGEwJV
-UzELMAkGA1UECBMCQ0ExFjAUBgNVBAcTDVNhbiBGcmFuY2lzY28xDTALBgNVBAoT
-BE1pc2sxDzANBgNVBAsTBkNsaWVudDEUMBIGA1UEAxMLbWlzay1jbGllbnQwHhcN
-MTgwMjIzMTE0MTQ4WhcNMjMwMTI4MTE0MTQ4WjBoMQswCQYDVQQGEwJVUzELMAkG
-A1UECBMCQ0ExFjAUBgNVBAcTDVNhbiBGcmFuY2lzY28xDTALBgNVBAoTBE1pc2sx
-DzANBgNVBAsTBkNsaWVudDEUMBIGA1UEAxMLbWlzay1jbGllbnQwggEiMA0GCSqG
-SIb3DQEBAQUAA4IBDwAwggEKAoIBAQDZynQhihI91W/040gjD5LRf1WpvvMSiHVO
-KP0ulpkVbhvVFb+2br0zL3eU9CQ7EFtc5v9DlKSArkczuGA5GD9dv8s8GDhQ4bED
-1bQWMq4rUhOnRW0jfcfKrCcQpsySzK0OkgXMfwGo11hOYnvFzlxUFlCKYd9qj0hq
-nZL08RVKe6maNO1V8tRGeuyC1U5oaVPsrj88gZj54S7Ape65EGkNcSWq+bu95NeD
-eVy4KfJtq1JmJEpTNmoXQthqV9q1uHY5aJOCwagPWl31DWCzDDjZ3szcj+cGIZKg
-cQXtjqpgysLxAeleNDiowP2WdYqbGBDBzfEMamyoSDJ8YPuZdBJFAgMBAAGjMjAw
-MB0GA1UdDgQWBBT5aiEsUHsZiA+jcF7et5cxXahyuzAPBgNVHREECDAGhwR/AAAB
-MA0GCSqGSIb3DQEBCwUAA4IBAQDDubv0I1/tDq4CsyiemQy9Tt+39t2buoI9ewM+
-I4HEbwlEdZ3sZAooUw+ZP5ZS7c1HPqjfXLQniaYkfoifc+f3m/YCSExt24ar6mF7
-jEEAkbTt2btg2oZSr/+19xVLKffAvWf9lvFu5GEyHbYSCFWE/3mtYnzlWbU+XiOH
-tBuytXr+63miohsGP2MUGfMenWD9PPMzcFC5bxbqjMR5tyTIfgDcn+bJdsV96lqq
-w6X/K9olgRs1Y+hXfDx3+ky+p3BuZxNCD8OUQH/hiEqc/0q1HW1MRja1V96z1ZLF
-Xy81jRh/P2u75taNQeIYZKyEBLy9uFfJmWVrkKmmMQBHTL/g
------END CERTIFICATE-----
