+++ settings.gradle
@@ -35,6 +35,7 @@ include ':misk-prometheus'
 include ':misk-redis'
 include ':misk-service'
 include ':misk-slack'
+include ':wisp-ssl'
 include ':misk-testing'
 include ':misk-transactional-jobqueue'
 include ':misk-zookeeper'
