+++ wisp-ssl/gradle.properties
@@ -1,0 +1,4 @@
+POM_ARTIFACT_ID=wisp-ssl
+POM_NAME=wisp-ssl
+POM_DESCRIPTION=SSL support for misk
+POM_PACKAGING=jar
