+++ wisp-ssl/src/main/kotlin/wisp/security/ssl/CertStore.kt
@@ -1,0 +1,6 @@
+package wisp.security.ssl
+
+import java.security.KeyStore
+
+/** A certificate and its private key. */
+data class CertStore(val keyStore: KeyStore)
