+++ wisp-ssl/src/main/kotlin/wisp/security/ssl/CertStoreConfig.kt
@@ -1,0 +1,7 @@
+package wisp.security.ssl
+
+data class CertStoreConfig constructor(
+  val resource: String,
+  val passphrase: String? = null,
+  val format: String = SslLoader.FORMAT_JCEKS
+)
