+++ wisp-ssl/src/main/kotlin/wisp/security/ssl/KeyStoreX509KeyManager.kt
@@ -1,4 +1,4 @@
-package misk.security.ssl
+package wisp.security.ssl
 
 import java.net.Socket
 import java.security.KeyStore
