+++ wisp-ssl/src/main/kotlin/wisp/security/ssl/PemComboFile.kt
@@ -1,4 +1,4 @@
-package misk.security.ssl
+package wisp.security.ssl
 
 import okio.Buffer
 import okio.BufferedSource
