+++ wisp-ssl/src/main/kotlin/wisp/security/ssl/TrustStore.kt
@@ -1,0 +1,6 @@
+package wisp.security.ssl
+
+import java.security.KeyStore
+
+/** A set of trusted root certificates. */
+data class TrustStore(val keyStore: KeyStore)
