+++ wisp-ssl/src/test/resources/ssl/client_cert_key_combo.pem
