+++ wisp-ssl/src/test/resources/ssl/client_rsa_cert_key_combo.pem
